// src/lib/shopify/generated/types.ts
export type CollectionsQuery = { collections: { nodes: Array<{ id: string, handle: string, title: string, image: { url: unknown, altText: string | null, width: number | null, height: number | null } | null }> } };

export type CollectionByHandleQueryVariables = Exact<{
  handle: string;
  country?: CountryCode | null | undefined;
  first?: number | null | undefined;
  after?: string | null | undefined;
}>;


export type CollectionByHandleQuery = { collection: { id: string, handle: string, title: string, description: string, image: { url: unknown, altText: string | null, width: number | null, height: number | null } | null, products: { pageInfo: { hasNextPage: boolean, hasPreviousPage: boolean, startCursor: string | null, endCursor: string | null }, nodes: Array<{ id: string, handle: string, title: string, productType: string, tags: Array<string>, availableForSale: boolean, priceRange: { minVariantPrice: { amount: unknown, currencyCode: CurrencyCode }, maxVariantPrice: { amount: unknown, currencyCode: CurrencyCode } }, featuredImage: { url: unknown, altText: string | null, width: number | null, height: number | null } | null, media: { nodes: Array<
            | { alt: string | null, previewImage: { url: unknown, width: number | null, height: number | null } | null }
            | { alt: string | null, previewImage: { url: unknown, width: number | null, height: number | null } | null }
            | { alt: string | null, previewImage: { url: unknown, width: number | null, height: number | null } | null }
            | { alt: string | null, previewImage: { url: unknown, width: number | null, height: number | null } | null }
          > }, options: Array<{ id: string, name: string, values: Array<string> }>, variants: { nodes: Array<{ id: string, title: string, availableForSale: boolean, quantityAvailable: number | null, selectedOptions: Array<{ name: string, value: string }>, price: { amount: unknown, currencyCode: CurrencyCode } }> } }> } } | null };